REFACTOR: Final da refatoração
