+++ CepApp/CepApp.App.csproj
@@ -69,6 +69,9 @@
 	  <MauiXaml Update="Views\Cep.xaml">
 	    <Generator>MSBuild:Compile</Generator>
 	  </MauiXaml>
+	  <MauiXaml Update="Views\Endereco.xaml">
+	    <Generator>MSBuild:Compile</Generator>
+	  </MauiXaml>
 	</ItemGroup>
 
 </Project>
