+++ CepApp/Views/MainPage.xaml
@@ -1,99 +1,14 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <TabbedPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-            xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
             UnselectedTabColor="Gray"
             SelectedTabColor="#512BD4"
             xmlns:android="clr-namespace:Microsoft.Maui.Controls.PlatformConfiguration.AndroidSpecific;assembly=Microsoft.Maui.Controls"
             android:TabbedPage.ToolbarPlacement="Bottom"
+            xmlns:views="clr-namespace:CepApp.Views"
             x:Class="CepApp.MainPage">
 
-    <ContentPage Title="CEP" IconImageSource="search_icon.svg" >
-        <ScrollView>
-            <VerticalStackLayout
-            Padding="30,0"
-            Margin="5"
-            VerticalOptions="Start">
-                <Border Padding="5">
-                    <Entry Keyboard="Numeric" Placeholder="Digite o CEP" Completed="BuscarCep"  VerticalOptions="Start" MaxLength="10">
-                        <Entry.Behaviors>
-                            <toolkit:MaskedBehavior  Mask="XX.XXX-XXX" />
-                        </Entry.Behaviors>
-                    </Entry>
-                </Border>
-
-                <Border x:Name="BorderError"  IsVisible="false">
-                    <Label x:Name="Error"/>
-                </Border>
-                <Label x:Name="Logradouro" Margin="0,15, 0, 0"/>
-
-                <Label x:Name="Bairro" Margin="0,15, 0, 0" />
-
-                <Label x:Name="Cidade" Margin="0,15, 0, 0"/>
-
-                <Label x:Name="Estado" Margin="0,15, 0, 0"/>
-
-                <Label x:Name="DDD" Margin="0,15, 0, 0"/>
-
-            </VerticalStackLayout>
-        </ScrollView>
-    </ContentPage>
-    <ContentPage Title="Endereço" IconImageSource="map_marker.svg">
-        <ScrollView>
-            <VerticalStackLayout
-                Padding="30,0"
-            Margin="5"
-            VerticalOptions="Start">
-
-                <Border Padding="5">
-                    <Picker x:Name="Ufs" Title="Selecione um estado..." SelectedIndexChanged="Ufs_SelectedIndexChanged"/>
-                </Border>
-
-                <Border  Padding="5" Margin="0,15,0,0">
-                    <Picker x:Name="Cidades" Title="Selecione uma cidade.."/>
-                </Border>
-
-                <Border x:Name="LogradouroBorders" Padding="5" Margin="0,15,0,0">
-                    <Entry x:Name="LogradouroRequest" Keyboard="Text" Placeholder="Digite o logradouro" Completed="LogradouroRequest_Completed"></Entry>
-                </Border>
-
-                <ListView x:Name="ListaCep" ItemsSource="{Binding ListCeps}" RowHeight="100" Margin="0,50">
-                    <ListView.ItemTemplate>
-                        <DataTemplate >
-                            <ViewCell >
-                                <Grid Padding="10,0,0,0" Margin="0,10,0,10">
-                                    <Grid.RowDefinitions>
-                                        <RowDefinition Height="Auto" />
-                                        <RowDefinition Height="Auto"/>
-                                        <RowDefinition Height="Auto"/>
-                                    </Grid.RowDefinitions>
-                                    <Grid.ColumnDefinitions>
-                                        <ColumnDefinition Width="300"/>
-                                    </Grid.ColumnDefinitions>
-                                    <Label 
-                                        Text="{Binding Cep}" 
-                                        FontAttributes="Bold"
-                                        FontSize="Medium"
-                                        />
-                                    <Label Grid.Row="1"
-                                        Text="{Binding Endereco}"
-                                        
-                                        LineBreakMode="CharacterWrap"
-                                        FontSize="Small"
-                                           />
-                                    <Label Grid.Row="2" 
-                                        Text="{Binding Detalhes}"
-                                        FontSize="Micro"
-                                           />
-                                </Grid>
-                            </ViewCell>
-                        </DataTemplate>
-                    </ListView.ItemTemplate>
-
-                </ListView>
-            </VerticalStackLayout>
-        </ScrollView>
-
-    </ContentPage>
+    
+    
 
 </TabbedPage>
